Added security and swagger docs and modified to constructor injection

+++ src/main/java/com/products/productDetails/controller/ProductController.java
@@ -2,11 +2,18 @@ package com.products.productDetails.controller;
 
 import com.products.productDetails.DTO.ProductDetailsDTO;
 import com.products.productDetails.DTO.ResponseObject.ProductDetailsRO;
-import com.products.productDetails.error.ProductNotFoundException;
+import com.products.productDetails.DTO.ResponseObject.ProductRO;
+import com.products.productDetails.DTO.ResponseObject.StockDetailsRO;
+import com.products.productDetails.exceptions.ProductNotFoundException;
 import com.products.productDetails.factory.ProductDetailsFactory;
 import com.products.productDetails.model.ProductDetails;
 import com.products.productDetails.service.ProductService;
 import com.products.productDetails.validator.ProductDetailsValidator;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -24,49 +31,85 @@ public class ProductController {
 
     private final ProductService productService;
 
-    private final ProductDetailsFactory productDetailsFactory;
-
     private final ProductDetailsValidator productDetailsValidator;
     private final Logger logger =
             LoggerFactory.getLogger(ProductController.class);
-    @PostMapping("/product")
+
+
+    @Operation(summary = "Create a Product")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Successfully created a product ",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ProductDetailsRO.class)) }),
+            @ApiResponse(responseCode = "400", description = "Invalid Data, Please Check Input Parameters and Value",
+                    content = @Content)
+    })
+    @PostMapping("/products")
     public ResponseEntity<ProductDetailsRO> createProduct
             (@Valid  @RequestBody ProductDetailsDTO productDetailsDTO){
         logger.info("Inside createProduct Method of ProductController Class");
-        ProductDetails productDetails = productDetailsFactory.mapFromProductDetailsDTO(productDetailsDTO);
-        ProductDetailsRO createdProduct = productDetailsFactory.mapFromProductDetails(
+        ProductDetails productDetails = ProductDetailsFactory.mapFromProductDetailsDTO(productDetailsDTO);
+        ProductDetailsRO createdProduct = ProductDetailsFactory.mapFromProductDetails(
                productService.createProduct(productDetails));
         return ResponseEntity.status(HttpStatus.CREATED).body(createdProduct);
     }
 
+    @Operation(summary = "Get all Product Details")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved all products details ",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ProductRO.class)) })
+    })
     @GetMapping("/products")
-    public ResponseEntity<List<ProductDetailsRO>> retrieveAllProducts(){
+    public ResponseEntity<List<ProductRO>> retrieveAllProducts(){
         logger.info("Inside retrieveAllProducts Method of ProductController Class");
         List<ProductDetails> productDetails = productService.retrieveAllProducts();
-        List<ProductDetailsRO> productDetailsRO = productDetails.stream()
-                .map(details->productDetailsFactory.mapFromProductDetails(details))
+        List<ProductRO> productRO = productDetails.stream()
+                .map(details->ProductDetailsFactory.mapFromProductDetailsDTO(details))
                 .toList();
-        return ResponseEntity.status(HttpStatus.OK).body(productDetailsRO);
+        return ResponseEntity.status(HttpStatus.OK).body(productRO);
     }
 
-    @GetMapping("/products/{id}")
-    public ResponseEntity<ProductDetailsRO> retrieveProductDetailsById(@PathVariable("id") long productId)
+    @Operation(summary = "Get Product Details based on productId")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved products details by given id",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ProductRO.class)) }),
+            @ApiResponse(responseCode = "400", description = "Invalid ProductId",
+                    content = @Content),
+            @ApiResponse(responseCode = "404", description = "No Product Found",
+                    content = @Content)
+
+    })
+    @GetMapping("/products/{productId}")
+    public ResponseEntity<ProductRO> retrieveProductDetailsById(@PathVariable("productId") long productId)
             throws ProductNotFoundException {
         logger.info("Inside retrieveProductDetailsById Method of ProductController Class");
-        ProductDetailsRO productDetailsRO = productDetailsFactory.mapFromProductDetails(
+        ProductRO productRO = ProductDetailsFactory.mapFromProductDetailsDTO(
                 productService.retrieveProductDetailsById(productId));
-        return ResponseEntity.status(HttpStatus.OK).body(productDetailsRO);
+        return ResponseEntity.status(HttpStatus.OK).body(productRO);
     }
 
-    @PutMapping("/products/{id}")
+    @Operation(summary = "Update Product based on productId")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Successfully created a product ",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ProductDetailsRO.class)) }),
+            @ApiResponse(responseCode = "400", description = "Invalid Data, Please Check Input Parameters and Value",
+                    content = @Content),
+            @ApiResponse(responseCode = "204", description = "Successfully updated the product details",
+                    content = @Content)
+
+    })
+    @PutMapping("/products/{productId}")
     public ResponseEntity<ProductDetailsRO> updateProductDetails
-            (@PathVariable("id") long productId,@Valid @RequestBody ProductDetailsDTO productDetailsDTO){
+            (@PathVariable("productId") long productId,@Valid @RequestBody ProductDetailsDTO productDetailsDTO){
         logger.info("Inside updateProductDetails Method of ProductController Class");
-        ProductDetails productDetails = productDetailsFactory.mapFromProductDetailsDTO(productDetailsDTO);
+        ProductDetails productDetails = ProductDetailsFactory.mapFromProductDetailsDTO(productDetailsDTO);
         Boolean isProductAvailable = productDetailsValidator.isProductExist(productId);
         if(!isProductAvailable){
             logger.info("There is no Product available with the given Id, Creating a new product");
-            ProductDetailsRO createdProduct = productDetailsFactory.mapFromProductDetails(
+            ProductDetailsRO createdProduct = ProductDetailsFactory.mapFromProductDetails(
                     productService.createProduct(productDetails));
             return ResponseEntity.status(HttpStatus.CREATED).body(createdProduct);
         }
@@ -74,11 +117,46 @@ public class ProductController {
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @DeleteMapping("/products/{id}")
-    public ResponseEntity<Void> removeProduct(@PathVariable("id") long productId){
+    @Operation(summary = "Delete Product based on productId")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Successfully deleted the product details",
+                    content = @Content),
+            @ApiResponse(responseCode = "404", description = "No Products available to Delete!",
+                    content = @Content)
+
+
+    })
+    @DeleteMapping("/products/{productId}")
+    public ResponseEntity<Void> removeProduct(@PathVariable("productId") long productId) throws ProductNotFoundException {
         logger.info("Inside removeProduct Method of ProductController Class");
+        Boolean isProductAvailable = productDetailsValidator.isProductExist(productId);
+        if(!isProductAvailable){
+            throw new ProductNotFoundException("No Products available to Delete!");
+        }
         productService.removeProduct(productId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+
+    @Operation(summary = "Get Stock Details based on productId")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved stock details by given id",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = StockDetailsRO.class)) }),
+            @ApiResponse(responseCode = "400", description = "Invalid ProductId",
+                    content = @Content),
+            @ApiResponse(responseCode = "404", description = "No Product Found",
+                    content = @Content)
+
+    })
+    @GetMapping("/stock-details/{productId}")
+    public ResponseEntity<StockDetailsRO> getStockInformation(@PathVariable("productId") long productId) throws ProductNotFoundException {
+        logger.info("Inside getStockInformation of ProductController Class");
+
+        ProductDetails productDetailsRO = productService.retrieveProductDetailsById(productId);
+
+        StockDetailsRO stockDetailsRO = ProductDetailsFactory.mapToStockDetails(productDetailsRO);
+
+        return ResponseEntity.status(HttpStatus.OK).body(stockDetailsRO);
     }
 
 }
